Fix error with PHP7 on method overloading
PHP7 is strictier on function overloading: they need to have the exact
same parameters as the parent function. In this case, we are lucky since
we do not need to replace the parent method, since the parent will never
call it. We can simply create a new method for our purpose.

Fixes #33